+++ CHANGELOG.md
@@ -1,5 +1,9 @@
-## [Unreleased]
+## 0.2.0
 
-## [0.1.0] - 2021-09-27
+- Uprade gem due [CVE-2021-41098](https://github.com/advisories/GHSA-2rr5-8q37-2w7h)
+- Add initializers config option, available initializers are [rails, sentry, sidekiq].
+- Move debug option to config.
+
+## 0.1.0
 
 - Test version for publish and test workflow.